% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp.R
\name{inpstring}
\alias{inpstring}
\title{Input line formatter}
\usage{
inpstring(..., cls = c("int", "dble", "char"), ns = NULL, ne = NULL)
}
\arguments{
\item{...}{objects to format}

\item{cls}{class of the input line}

\item{ns}{integer; the beginning index}

\item{ne}{integer; the ending index}
}
\description{
Input line formatter
}
\details{
The user should not need to use this function.
}
\examples{
inpstring(a=1)
inpstring(a=1,b=1)
}
\seealso{
Other Input.file: \code{\link{generator}},
  \code{\link{inp}}
}

