% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lith.R
\name{lithology}
\alias{lithology}
\alias{plot.lith}
\title{Lith}
\usage{
lithology(Depth, ShearModulus = 1e+09, Skempton = 0.85, Diffusivity = 0.1,
  Nu = 0.25, Nuu = 0.35, Number = NULL)

\method{plot}{lith}(L, no.layout = TRUE, ...)
}
\description{
Lith
}
\examples{
# just do something random to demonstrate
n <- 5 # five layers
deps <- seq_len(n)-1 # at these depths
rand <- runif(n)
(l <- lithology(deps, rand, rand, rand))
}

