% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp-loader.R
\name{print.poel-inp}
\alias{print.poel-inp}
\title{Methods for 'poel-inp' class}
\usage{
\method{print}{poel-inp}(x)
}
\description{
Methods for 'poel-inp' class
}
\seealso{
\code{\link{read_inp}} \code{\link{read_t}}
}

