% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{poel-params}
\alias{poel-params}
\alias{poel_alpha}
\alias{poel_beta}
\alias{poel_chi}
\alias{poel_lambda}
\title{Parameter sensitivity to Poisson's ratio}
\usage{
poel_lambda(nu)

poel_alpha(nu, nuu)

poel_beta(nu, nuu)

poel_chi(nu, nuu)
}
\arguments{
\item{nu}{numeric; Poisson's ratio}

\item{nuu}{numeric; undrained Poisson's ratio}
}
\description{
Functions to investigate sensitivity of normalized-parameters
}
\details{
\code{\link{poel_lambda}} gives the Lame constant, normalized by
\eqn{\mu}, the elastic shear modulus; this has no dependence on the undrained
Poisson's ratio

\code{\link{poel_alpha}} gives the effective stress coefficient, normalized by
\eqn{1 / B} where B is Skempton's coefficient

\code{\link{poel_beta}} gives the bulk compressibility, normalized by
\eqn{1 / \mu / B^2}

\code{\link{poel_chi}} gives the Darcy conductivity, normalized by
\eqn{D / \mu / B^2} where D is the hydraulic diffusivity, which is proportional
to permeability
}
\examples{
a1 <- poel_alpha(0.25,0.33)
a2 <- poel_alpha(0.20,0.40)
a2/a1

# try multiple values
poel_alpha(0.25,seq(0,1,by=0.1))
}

