% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-loader.R
\name{print.poel-t}
\alias{print.poel-t}
\title{Methods for 'poel-t' class}
\usage{
\method{print}{poel-t}(x, n = 100)
}
\arguments{
\item{n}{integer; the maximum number of indices to print}
}
\description{
Methods for 'poel-t' class
}
\seealso{
\code{\link{read_t}} \code{\link{read_inp}}
}
