% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp-loader.R
\name{read_inp}
\alias{read_inp}
\title{Read .inp files}
\usage{
read_inp(fi, version = 2012)
}
\arguments{
\item{fi}{character; full-path to inp file}

\item{version}{the poel-version \code{fi} is formatted for; coerced to character}
}
\description{
Read-in and parse .inp files -- input files for poel -- for later use
}
\examples{

# Example .inp #1
inpfi1 <- system.file('pe12_dk_s0_b1.inp', package='poel')
res1 <- read_inp(inpfi1)
print(res1)

# Example .inp #2
inpfi2 <- system.file('pe12_dk_s0_b2.inp', package='poel')
res2 <- read_inp(inpfi2)
print(res2)

}
\seealso{
\code{\link{read_t}}
}

