% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-loader.R
\name{read_t}
\alias{read_t}
\title{Read .t files -- poel-result files}
\usage{
read_t(fi, version = 2012)
}
\arguments{
\item{fi}{character; full-path to inp file}

\item{version}{the poel-version \code{fi} is formatted for; coerced to character}
}
\description{
Read-in and maniputate .t files, the standard format used in poel output
}
\examples{
tfi <- system.file('pp.t.gz', package='poel')
res <- read_t(tfi)
print(res)
}
\seealso{
\code{\link{read_inp}}
}
