% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stepvec}
\alias{stepvec}
\title{Convert vector into step-like vector}
\usage{
stepvec(x, y)
}
\arguments{
\item{x, y}{numeric vectors giving the coordinates of the points to be
converted into a step-like function}
}
\description{
Take a vector of values (with some set of indices)
and create a step-like function by shifting indices.
}
\details{
This will be useful for creating source-time-functions
from realistic data, that are formatted appropriately for poel.
}
\examples{

SV <- stepvec(seq_len(10), rnorm(10))
SV

\dontrun{
plot(y ~ x, SV, type='b')
}
}
