% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generator}
\alias{generator}
\alias{generator.default}
\alias{generator.inp}
\title{Generate an input file suitable to \code{poel}}
\usage{
generator(x, ...)

\method{generator}{inp}(x, ...)

\method{generator}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters}
}
\description{
Generate an input file suitable to \code{poel}
}
\seealso{
Other Input.file: \code{\link{inpstring}};
  \code{\link{inp}}, \code{\link{inp.default}},
  \code{\link{inp.lith}}
}

