\name{generator}
\alias{generator}
\title{Generate a .inp file}
\usage{
  generator(x, ...)
}
\arguments{
  \item{x}{object}

  \item{...}{additional parameters}
}
\description{
  Generate a .inp file
}
\seealso{
  Other Input.file: \code{\link{.inpstring}},
  \code{\link{inp}}, \code{\link{inp.default}},
  \code{\link{inp.lith}}
}

