% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lith.R
\name{lithology}
\alias{depth_param_functions}
\alias{depth_param_functions.lith}
\alias{is.lith}
\alias{lith}
\alias{lithology}
\alias{plot.lith}
\title{Poel's lithology structure}
\usage{
lithology(Depth, ShearModulus = 1e+09, Skempton = 0.85, Diffusivity = 0.1,
  Nu = 0.25, Nuu = 0.35, Number = NULL)

is.lith(x, ...)

\method{plot}{lith}(L, no.layout = TRUE, add = FALSE, plot.inds = c(3, 4,
  5, 6, 7), ...)

depth_param_functions(x, ...)

\method{depth_param_functions}{lith}(x, ...)
}
\description{
Manipulate or generate poel-type lithology specifications
}
\examples{

# just do something random to demonstrate
set.seed(1234)

n <- 5 # five layers
deps <- seq_len(n)-1 # at these depths
rand <- runif(n)

(l <- lithology(deps, rand, rand, rand))
is.lith(l)

# plot values
plot(l)

# Functions for plotting lith params at any depth:
funs <- depth_param_functions(l)

HD <- funs[['Diffusivity']]
# get many more values than there were before
new.depths <- sort(jitter(rep(deps, each=n)))
new.hd <- HD(new.depths)
plot(new.depths, new.hd, type='l')

}

