% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{poel-params}
\alias{poel-params}
\alias{poel_lambda}
\alias{poel_alpha}
\alias{poel_beta}
\alias{poel_chi}
\title{Parameter sensitivity to Poisson's ratio}
\usage{
poel_lambda(nu)

poel_alpha(nu, nuu)

poel_beta(nu, nuu)

poel_chi(nu, nuu)
}
\arguments{
\item{nu}{numeric; Poisson's ratio}

\item{nuu}{numeric; undrained Poisson's ratio}
}
\description{
Functions to investigate sensitivity of normalized-parameters
}
\details{
\code{\link{poel_lambda}} gives the Lame constant, normalized by
\eqn{\mu}, the elastic shear modulus; this has no dependence on the undrained
Poisson's ratio

\code{\link{poel_alpha}} gives the effective stress coefficient, normalized by
\eqn{1 / B} where \eqn{B} is Skempton's coefficient. Hence, to obtain Biot's pore
pressure coefficient (generally written as \eqn{\alpha}) take this value and
divide by \eqn{B}.

\code{\link{poel_beta}} gives the bulk compressibility, normalized by
\eqn{1 / \mu / B^2}

\code{\link{poel_chi}} gives the Darcy conductivity, normalized by
\eqn{D / \mu / B^2} where D is the hydraulic diffusivity, which is proportional
to permeability
}
\examples{
a1 <- poel_alpha(0.25,0.33)
a2 <- poel_alpha(0.20,0.40)
a2/a1

# try multiple values
poel_alpha(0.25,seq(0,1,by=0.1))

# Make a sensitivity matrix
# Poisson's ratios
nu <- nuu <- seq(0,0.5,by=0.01)
alpgrd <- outer(nu, nuu, "poel_alpha")
B <- 0.75 # Skempton's coefficient
biot <- alpgrd/B

fields::image.plot(nu, nuu, biot, asp=1)
contour(nu, nuu, biot, add=TRUE)

}
\references{
H.-J. Kümpel; Poroelasticity: parameters reviewed. Geophys J Int 1991; 105 (3): 783-799. doi: 10.1111/j.1365-246X.1991.tb00813.x
}
